style: change register route
Change the register route:
backController->confirmRegister($id, $token)
=> frontController->confirmRegister($email, $token);

issue #45
