+++ App/template/front/register.php
