+++ App/template/mail/mail_confirmation.php
@@ -14,7 +14,7 @@
         <p>Bonjour <?= $pseudo ?>!</p>
         <p>
             Merci de vous êtes inscrit ;)</br>
-            Afin de confirmer votre inscription veuillez cliquer sur le lien suivant: <a href="http://localhost/ocr/blog/index.php?route=email-confirmation&id=<?= $id; ?>&token=<?= $token; ?>" alt="">Valider votre compte</a>
+            Afin de confirmer votre inscription veuillez cliquer sur le lien suivant: <a href="http://localhost/ocr/blog/index.php?route=inscription&action=confirmation&email=<?= $email; ?>&token=<?= $token; ?>" alt="">Valider votre compte</a>
         </p>        
 
     </body>
